+++ lib/sg_pt_osf1.c
@@ -249,7 +249,7 @@ void set_scsi_pt_packet_id(struct sg_pt_base * vp, int pack_id)
 {
 }
 
-void set_scsi_pt_tag(struct sg_pt_base * vp, unsigned long long tag)
+void set_scsi_pt_tag(struct sg_pt_base * vp, uint64_t tag)
 {
     struct sg_pt_osf1_scsi * ptp = &vp->impl;
